# MAINT: Spherical Voronoi sort simplify