* `sort_vertices_of_regions()` was simplied to:
  * only a single call location for `remaining_filter()`
  * remove a spurious `break` check on `cs_identified`
  * with the above check removed, `cs_identified` can be removed (no longer used)
  * removed extra `sorted_vertices` and `remaining_count` lines by placing them strategically in the `while` loop
  * the `while` loop was then replaced with a `for` loop to avoid need for the silly `remaining_count` variable
  * another spurious `break` statement was removed
  * avoided unncessary usage of `ARRAY_FILLER` in `sorted_vertices` given that simple slicing of `sorted_vertices` can be used to deal with polygons with different edge counts
  * these simplifications then allowed me to determine that the algorithm was spending many lines of code dealing with this jagged polygon data structure situation, allowing for a dramatic simplification, removing the undesirable `DEF` constant and the `remaining_filter()` helper function thus became pointless

[skip circle] [skip cirrus]

~I was looking forward to trying to use `asv` to check performance, but either the latest version or some combination of build system modernizations seems to be catching me off guard with `mesonpy`-related errors when using `asv continuous` to compare two hashes/branches (I use `virtualenv` as the backend in the config file).~