@tylerjereddy 2023-09-11
`asv` benchmark results on x86_64 Linux:

`asv continuous -e -b ".*SphericalVor.*" main treddy_sph_vor_sort_simplify`

```
| Change   | Before [9d5d4f86] <main>   | After [603d021b] <treddy_sph_vor_sort_simplify>   |   Ratio | Benchmark (Parameter)                                                       |
|----------|----------------------------|---------------------------------------------------|---------|-----------------------------------------------------------------------------|
| -        | 60.8±2μs                   | 53.5±1μs                                          |    0.88 | spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(10, 3)    |
| -        | 358±4μs                    | 282±2μs                                           |    0.79 | spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(100, 3)   |
| -        | 3.27±0.05ms                | 2.53±0.01ms                                       |    0.77 | spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(1000, 3)  |
| -        | 16.2±0.1ms                 | 12.4±0.03ms                                       |    0.77 | spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(5000, 3)  |
| -        | 32.4±0.5ms                 | 24.7±0.2ms                                        |    0.76 | spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(10000, 3) |
| -        | 197±3μs                    | 124±0.6μs                                         |    0.63 | spatial.SphericalVorSort.time_spherical_polygon_vertex_sorting(100)         |
| -        | 2.00±0.03ms                | 1.25±0ms                                          |    0.63 | spatial.SphericalVorSort.time_spherical_polygon_vertex_sorting(1000)        |
| -        | 20.3±0.3ms                 | 12.7±0.06ms                                       |    0.63 | spatial.SphericalVorSort.time_spherical_polygon_vertex_sorting(10000)       |
| -        | 10.2±0.2ms                 | 6.33±0.02ms                                       |    0.62 | spatial.SphericalVorSort.time_spherical_polygon_vertex_sorting(5000)        |
| -        | 20.0±0.9μs                 | 12.2±0.05μs                                       |    0.61 | spatial.SphericalVorSort.time_spherical_polygon_vertex_sorting(10)          |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```